Subject: Changes to the tool_interface Description: Changes to the tool_interface
Added a few more functions that I needed for ccdb work.

Note that this interface is by no means comprehensive yet.

Created new header files which are meant to be the publicly visible API.
This hides a few implementation details found in the alps_*.h files and
only exposes what is meant to be shown.

Fixed a bug in ld_val where if the LD_AUDIT environment variable was not
defined, it would segfault on strdup due to a null pointer.
Inspected by: rwm
Review info:
submitted by andrewg on 03/10/11 16:25:17 CST
reviewer information set by andrewg on 03/10/11 16:25:17 CST  =>rwm
accepted by rwm on 03/10/11 18:41:21