Subject: Bugfix for CTI libraries Description: Bugfix for CTI libraries
Fixed a bug where I was iterating over an array of strings, and
was simply checking for the pointer into this array to be a null pointer. This
is invalid in the case were I have a full list and I suddenly point off the
end of the array of strings. Memory is not always guaranteed to be null.
Inspected by: rwm
Review info:
submitted by andrewg on 07/27/11 15:36:19 CDT
reviewer information set by andrewg on 07/27/11 15:36:19 CDT  =>rwm
accepted by rwm on 2014-11-17 15:38:08 CST