Subject: Cray Tool Interface bugfix Description: Cray Tool Interface bugfix
Fixed a bug in the getAppHostsPlacement function that would incorrectly
report the number of PEs on a Nid when there was only one Nid in the app.
Previously if we only found one Nid in the app, we would only return one
PE. This wasn't an issue for ccdb with MRNet since we still had the hostname.
However SCI relies heavily on the number of PEs on the Nid and broke when
there was only a single Nid. This issue is fixed with this commit.
Inspected by: rwm
Review info:
submitted by andrewg on 10/27/11 16:42:44 CDT
reviewer information set by andrewg on 10/27/11 16:42:44 CDT  =>rwm
accepted by rwm on 10/27/11 19:32:34