The result of this mod is that sites such as CSCS that run 'srun' from a whitebox are now supported.
This mod allows CTI to run on nodes, whether or not the
file /proc/cray_xt/nid exists. If it does exist it is used,
but if not, there is a fallback to libc gethostname.  CTI is
no longer dependent on the existance of a NID.

There were changes to the 'fe', 'be', and 'dl' layers.

Note that *cray_slurm* routines are clones of the *slurm*
routines and will collapse down to the *slurm* version when
we merge those two workload managers in a future mod.

Details:

  scripts/release_info_sles: Updates with a description of
    this mod.
  release_versioning: Incremented revision for BE and FE.

  src/frontend/cray_slurm_fe.c:
    Changed _cti_cray_slurm_getHostName to have a fallback to
      use gethostname() if /proc/cray_xt/nid does not exist.
  src/frontend/slurm_fe.c:
    Changed _cti_slurm_getHostName to have a fallback to
      use gethostname() if /proc/cray_xt/nid does not exist.

  src/backend/cray_slurm_be.c:
    Remove dealing with nodes by their NID and switched to their hostname
      (This is just like Alex did for CS platforms.)
    Changed _cti_be_cray_slurm_getNodeHostname to have a fallback to
      use gethostname() if /proc/cray_xt/nid does not exist.
  src/backend/slurm_be.c:
    Changed _cti_be_slurm_getNodeHostname to have a fallback to
      use gethostname() if /proc/cray_xt/nid does not exist.

  src/daemon/cray_slurm_dl.c:
    Changed _cti_cray_slurm_getNodeID to have a fallback to
      use a hash of gethostname() if /proc/cray_xt/nid does not exist.
  src/daemon/slurm_dl.c:
    Changed _cti_slurm_getNodeID to have a fallback to
      use a hash of gethostname() if /proc/cray_xt/nid does not exist.

Bug Comment: The result of this mod is that sites such as CSCS that
run 'srun' from a whitebox are now supported.\comment
Bugs Affected: 849773
Bug Action: RESOLVE