Fix for PE-25075
Major rework of entire interface. Use a proper singleton object for the
frontend that is thread safe. Ensure that the iface always goes through
the FE singleton to ensure construction on first use.

New ownership model to guarantee cleanup. A FE owns all Apps who own all
Sessions who own all Manifests. When one of these is destructed, all
owned objects underneath also get destructed. Each owned object has a
weak_ptr (or reference in the case of an App to the FE since that is not
a shared_ptr) to its owning object. That way owned objects can query
information stored in the owning object.

The iface has an independent map from an id to a weak_ptr. If the
weak_ptr goes out of scope, it is automatically removed from the map and
proper error returned via the iface.