Fix for PE-26486
Sanity check the CTI_BASE_DIR_ENV_VAR setting. Otherwise fall back to
use the hard coded install locations. For relocatable installs, we want to
allow the env var to be used as an override.

I also opened PE-26542 so that we can make this check stronger. Relying
on a blind env var is very dangerous. We can make it less dangerous by
using a checksum on say the be_daemon binary for sanity.